refactor: extract createExportRapportTool into tools/exportRapport.ts
